(CI): update gh pages script

+++ scripts/deploy-to-gh-pages-ci.sh
@@ -9,11 +9,19 @@ SHA=`git rev-parse --verify HEAD`
 # config git
 git config --global user.email "travis@travis-ci.com"
 git config --global user.name "Travis CI"
+
 # deploy
 npm run build:web
 cd web-dist
 git init
 git add .
 git commit -m "Deploy to GitHub Pages: ${SHA}"
+
+# If there are no changes to the compiled out (e.g. this is a README update) then just bail.
+if [ -z `git diff --exit-code` ]; then
+    echo "No changes to the output on this push; exiting."
+    exit 0
+fi
+
 git push --force --quiet "https://${GITHUB_TOKEN}@github.com/${TRAVIS_REPO_SLUG}.git" master:gh-pages > /dev/null 2>&1
 echo "successfully deployed commit ${SHA} to branch gh-pages"
